+++ internal/quipple/compiler/dag_nodes.go
