+++ internal/quipple/preprocess/registry.go
