+++ internal/quipple/preprocess/shorthand.go
